+++ core/CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Locking the device by holding finger on the homescreen for 2.5 seconds.  [#1404]
 - Public key to ECDHSessionKey.  [#1518]
+- Decred staking.  [#1249]
 
 ### Changed